First commit
Here's an initial sketch of the e-g library, or a set of macros to
define new types based on concrete values of their inhabitants.

This project is mostly for fun, but can be useful in enforcing
boundaries across codebases. It allows the users of libraries to define
their own "extension types" that refer to the members of some other type
through exposed functions. This ensures the type definitions that are
dependent on the types in another library are automatically updated
should those types change.

In this sense they are minimally dependent types insofar as they depend
on other type definitions and functions.